Revert "Merge pull request #94 from Icinga/move-form-validation-to-formelements"
This reverts commit 95207299437ccce4a1a3981ee3f3d70152c5fb2c, reversing
changes made to c3293ceea93a2a2f4ecd4c2860e1e8aef53ac092.
